refactor: restructure i18n system with better organization
- Move from locales/ to i18n/translations/ structure
- Improve naming convention (pt-br/ → pt/, en-us/ → en/)
- Update i18n configuration in i18n/index.js
- Remove old locales/ and 118n/ systems
