--- webapp/frontend/src/locales/index.js
@@ -1,6 +1,0 @@
-import en from './en';
-import pt from './pt';
-
-export { en, pt };
-
-export default { en, pt };
